+++ public/js/audio.js
@@ -1,3 +1,8 @@
+if ('serviceWorker' in navigator) {
+  navigator.serviceWorker.register('/sw_cache.js')
+    .catch(console.error);
+}
+
 const audioArr = [
   new Audio('audio/1.mp3'),
   new Audio('audio/2.mp3'),