+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- caching
+- service worker for caching
+- viewport settings
 
 ## [0.3.0] - 2020-11-04
+### Added
 - add `node_modules` and `.idea` to `.gitignore` file
 - add audio files
 - ability to run audio files